Switch from imagemin-power-cli to imagemin-cli
Adrian had issues running imagemin-power-cli and it was a bit of a hack anyway to provide recursive processing through `/dist` dir. Now uses main imagemin-cli and `find` command to run it manually on each folder.

The command will be simplified in the future when imagemin finally releases recursive folder support.

https://github.com/imagemin/imagemin/issues/87
https://github.com/imagemin/imagemin-cli/pull/11